Only alter the last two bytes of the MAC address
Xlink Kia uses the first 3 bytes of the MAC address to determin if it
should connect it, so keep those constant.

The other 3 bytes can be any value. only 2 of these 3 are set based on
the host PC as it appears Xlink only uses the last two bytes to generate
a IP address.